restricted pages

+++ django1/base/views.py
@@ -1,6 +1,8 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
@@ -15,6 +17,9 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == "POST":
         userName = request.POST.get('username')
@@ -53,11 +58,13 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room' : room}
     return render(request, 'base/room.html', context)
-    
+
+@login_required(login_url='/login')   
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -70,9 +77,14 @@ def createRoom(request):
 
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not allowed')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -83,8 +95,15 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Not allowed')
+
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
